Alliance v0.5.5.0 - PvC fixes (#106)
# Alliance v0.5.5.0 - Gameplay, UI and Spawn menu improvements

## Summary
This release fixes multiple issues related to player spawn menu, banner
renders and server settings. It adds the ability to customize agent
health and also improve the build system.

## Features
### Build System
- Prefabs containing `AL_TriggerAction` scripts can now be placed
dynamically during a mission using the Build System
- Added example "autospawn" prefabs
- Prefabs can now be deleted while in spectator mode

### Server options
- Added categories and dependencies to game options to improve
readability
- Options are now automatically hidden when their parent option is
disabled
- Categories can be collapsed and expanded
- Removed several unused options

### Gameplay
- Added native maps to the CVC map roster
- Health multipliers from selected characters are now correctly applied
in Lobby, Battle, PvC, CvC and respawn systems

### UI
- The formation system indicator is now hidden when the system is
disabled
- Removed duplicate entries in the character editor of the player spawn
menu
- Improved MPCharacter loading and availability in both game and Modding
Kit contexts
- Changed the Scenario Editor shortcut from `CTRL+O` to `CTRL+P` to
avoid conflicts

## Fixes
- Fixed commander camera and commander status detection when commanders
are playing in PvC
- Fixed native sounds not playing due to missing references in the
MBProj file
- Fixed a crash when an agent attempted to capture an objective without
belonging to a valid side
- Fixed a crash caused by invalid election timer compression settings
- Fixed a crash when launching a server with a warmup duration lower
than 60 seconds
- Fixed banner recoloring issues during team creation
- Fixed banner comparison failures caused by floating-point precision
errors
- Fixed banner rendering in multiple UI
- Fixed a bug locking the character editor of Player Spawn menu by
adding a default culture to formations

## Content Changes
- Removed unnecessary SceneEditData files
- Player Spawn menu now only display characters linked to valid
MPClassDivisions


---------

Co-authors: @spawnpbbw & @Byak0

+++ Alliance.Common/Extensions/PlayerSpawn/NetworkMessages/PlayerSpawnMenuMsg.cs
@@ -46,13 +46,13 @@ namespace Alliance.Common.Extensions.PlayerSpawn.NetworkMessages
 		}
 
 		// From server - Sends the election timer to a specific peer.
-		public static void SendElectionStatusToPeer(bool enable, float timer, NetworkCommunicator peer)
+		public static void SendElectionStatusToPeer(bool enable, int timer, NetworkCommunicator peer)
 		{
 			SendToClient(new SetElectionStatus(enable, timer), peer);
 		}
 
 		// From server - Broadcasts the election timer to all players in the game.
-		public static void SendElectionStatusToAll(bool enable, float timer = 0f)
+		public static void SendElectionStatusToAll(bool enable, int timer = 0)
 		{
 			SendToClient(new SetElectionStatus(enable, timer));
 		}
